+++ src/watcher/models/notifier/index.js
@@ -1,0 +1,8 @@
+import Notifier from './base';
+import polymorphs from './polymorphs';
+
+export const notifierPool = {};
+export {
+  Notifier,
+  polymorphs as notifierTypes,
+};
