+++ src/watcher/models/notifier/polymorphs/index.js
@@ -1,0 +1,4 @@
+import SlackNotifier from './slack_notifier';
+import AwsSnsNotifier from './aws_sns_notifier';
+
+export default { SlackNotifier, AwsSnsNotifier };
